Add addToHistory unit tests
