--- src/extension-test/unit/repl-history-test.ts
@@ -1,12 +1,0 @@
-import { expect } from 'chai';
-import * as replHistory from '../../results-output/repl-history';
-
-describe('repl-history', () => {
-    describe('addToHistory', () => {
-        it('push text to history array', () => {
-            const history = [];
-            const newHistory = replHistory.addToHistory(history, 'hello');
-            expect(newHistory[0]).equal('hello');
-        });
-    });
-});
